Move api/stdcopy to api/pkg/stdcopy
